refactor(renderer): group agent and profile ownership
