+++ apps/desktop/src/renderer/App.tsx
@@ -1,19 +1,19 @@
-import { AgentPanel } from './components/AgentPanel'
+import { AgentPanel } from './agents/chat/AgentPanel'
 import { CenterWorkspace, type JobListingRequest } from './components/CenterWorkspace'
 import { JobNavigator } from './components/JobNavigator'
 import { SettingsPanel } from './components/SettingsPanel'
 import { StatusBar } from './components/StatusBar'
 import { WorkbenchLayout } from './components/WorkbenchLayout'
-import { CareerProfileWorkspace } from './components/CareerProfileWorkspace'
-import { hasCachedCareerProfile } from './hooks/useCareerProfile'
+import { CareerProfileWorkspace } from './career-profile/CareerProfileWorkspace'
+import { hasCachedCareerProfile } from './career-profile/work-arrangement/useCareerProfile'
 import { WorkspaceBar, type WorkspaceBarWorkspace } from './components/WorkspaceBar'
 import { BrowseWorkspace } from './components/BrowseWorkspace'
-import { useAgentAvatarPreference } from './agent-avatar/useAgentAvatarPreference'
+import { useAgentAvatarPreference } from './agents/avatar/useAgentAvatarPreference'
 import { DocxDocumentEditorShell } from './document-editor/DocxDocumentEditorShell'
 import { useConnectivity } from './hooks/useConnectivity'
 import { useJobs } from './hooks/useJobs'
-import { useAgentSessions } from './hooks/useAgentSessions'
-import { useConnectedAgents } from './hooks/useConnectedAgents'
+import { useAgentSessions } from './agents/chat/useAgentSessions'
+import { useConnectedAgents } from './agents/connected-agents/useConnectedAgents'
 import { useWorkspace } from './hooks/useWorkspace'
 import { useTheme } from './theme/useTheme'
 import { useCallback, useEffect, useRef, useState } from 'react'
@@ -26,7 +26,7 @@ import type {
   SetupSnapshot
 } from '../shared/contracts'
 import { OnboardingScreen } from './onboarding/OnboardingScreen'
-import { NewAgentChatDialog } from './components/NewAgentChatDialog'
+import { NewAgentChatDialog } from './agents/new-chat/NewAgentChatDialog'
 
 export function App() {
   const [setup, setSetup] = useState<SetupSnapshot | null>(
