+++ apps/desktop/src/renderer/agents/avatar/AgentAvatar.test.tsx
