+++ apps/desktop/src/renderer/agents/avatar/AgentAvatar.tsx
