+++ apps/desktop/src/renderer/agents/avatar/agentAvatars.ts
