+++ apps/desktop/src/renderer/agents/avatar/assets/crimson.webp
