+++ apps/desktop/src/renderer/agents/avatar/assets/ember.webp
