+++ apps/desktop/src/renderer/agents/avatar/assets/forest.webp
