+++ apps/desktop/src/renderer/agents/avatar/assets/ivory.webp
