+++ apps/desktop/src/renderer/agents/avatar/assets/midnight.webp
