+++ apps/desktop/src/renderer/agents/avatar/assets/mint.webp
