+++ apps/desktop/src/renderer/agents/avatar/assets/ninja.webp
