+++ apps/desktop/src/renderer/agents/avatar/assets/phantom.webp
