+++ apps/desktop/src/renderer/agents/avatar/assets/shadow.webp
