+++ apps/desktop/src/renderer/agents/avatar/assets/starlight.webp
