+++ apps/desktop/src/renderer/agents/avatar/assets/venom.webp
