+++ apps/desktop/src/renderer/agents/avatar/useAgentAvatarPreference.test.tsx
