+++ apps/desktop/src/renderer/agents/avatar/useAgentAvatarPreference.ts
