+++ apps/desktop/src/renderer/agents/chat/AgentPanel.tsx
@@ -1,11 +1,11 @@
 import { ArrowDown, BriefcaseBusiness, CircleAlert, LoaderCircle, RotateCcw, Send, Square, UserRound, WifiOff } from 'lucide-react'
 import { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react'
 
-import type { ConnectivityState } from '../../shared/contracts'
-import { AgentAvatar } from '../agent-avatar/AgentAvatar'
-import type { AgentAvatarId } from '../agent-avatar/agentAvatars'
-import type { AgentSessionsController } from '../hooks/useAgentSessions'
-import { initialAgentConversationState } from '../hooks/useAgentConversation'
+import type { ConnectivityState } from '../../../shared/contracts'
+import { AgentAvatar } from '../avatar/AgentAvatar'
+import type { AgentAvatarId } from '../avatar/agentAvatars'
+import type { AgentSessionsController } from './useAgentSessions'
+import { initialAgentConversationState } from './useAgentConversation'
 import { AgentActivityGroup } from './AgentActivityGroup'
 import { ActivityRow } from './ActivityRow'
 import { AssistantMarkdown } from './AssistantMarkdown'
