+++ apps/desktop/src/renderer/installation-profiles/InstallationProfileMenu.test.tsx
