+++ apps/desktop/src/renderer/installation-profiles/InstallationProfileMenu.tsx
